Web Worker Examples
Got an embedded web worker in the html working, trying out examples
that use external files but they seem to be having a hard time finding
the script for the worker for some reason. Will look more in the
morning. Also found out why the sound stuff updates in the background
despite not being hooked up to threading yet.